Create README.md
I'm a passionate coder currently diving into the world of Data Science. With a strong foundation in Python and advanced SQL, they are actively honing their skills in Power BI and Excel to become a well-rounded data specialist.

Enthusiastic and collaborative, also I am eager to contribute to projects and work alongside others to learn and grow in the data science field.